push

+++ src/tic_tac_toe_simulator.rs
@@ -47,7 +47,6 @@ impl<'a> GameSimulator<'a>{
             context: ctx
         }
     }
-    #[inline(always)]
     fn check_horizontally(&self,x: u8, y: u8,to_check: XO) -> bool{
         let mut count = 0u8;
         while let Some(gotten_y) = self.board.get(y as usize)
@@ -62,7 +61,6 @@ impl<'a> GameSimulator<'a>{
         }
         count == Self::NUMBER_TO_WIN
     }
-    #[inline(always)]
     fn check_vertically(&self,x: u8, y: u8,to_check: XO) -> bool{
         let mut count = 0u8;
         while let Some(gotten_y) = self.board.get((y + count) as usize)
@@ -77,7 +75,6 @@ impl<'a> GameSimulator<'a>{
         }
         count == Self::NUMBER_TO_WIN
     }
-    #[inline(always)]
     fn check_diagonally_forward(&self,x: u8, y: u8,to_check: XO) -> bool{
         let mut count = 0u8;
         while let Some(gotten_y) = self.board.get((y + count) as usize)
@@ -92,7 +89,6 @@ impl<'a> GameSimulator<'a>{
         }
         count == Self::NUMBER_TO_WIN
     }
-    #[inline(always)]
     fn check_diagonally_backward(&self,x: u8, y: u8,to_check: XO) -> bool{
         let mut count = 0u8;
         while let Some(gotten_y) = self.board.get((y + count) as usize)
@@ -107,18 +103,17 @@ impl<'a> GameSimulator<'a>{
         }
         count == Self::NUMBER_TO_WIN
     }
-    #[inline(always)]
     fn check_for_winner(&self) -> Option<XO>{
         for to_check in [XO::X, XO::O]{
             for y in 0..self.area{
                 for x in 0..self.area{
-                    if self.check_horizontally(x, y, to_check){
+                    if self.check_horizontally(x as u8,y as u8,to_check){
                         return Some(to_check)
-                    } else if self.check_vertically(x, y, to_check){
+                    } else if self.check_vertically(x as u8,y as u8,to_check){
                         return Some(to_check)
-                    } else if self.check_diagonally_forward(x, y, to_check){
+                    } else if self.check_diagonally_forward(x as u8,y as u8,to_check){
                         return Some(to_check)
-                    } else if self.check_diagonally_backward(x, y, to_check){
+                    } else if self.check_diagonally_backward(x as u8,y as u8,to_check){
                         return Some(to_check)
                     }
                 }
